+++ praw/models/reddit/comment.py
@@ -196,11 +196,6 @@ class Comment(InboxableMixin, UserContentMixin, FullnameMixin, RedditBase):
         self.__dict__.update(other.__dict__)
         self._fetched = True
 
-    def _fetch_data(self):
-        name, fields, params = self._fetch_info()
-        path = API_PATH[name].format(**fields)
-        return self._reddit.request(method="GET", params=params, path=path)
-
     def _fetch_info(self):
         return "info", {}, {"id": self.fullname}