+++ praw/models/reddit/redditor.py
@@ -182,11 +182,6 @@ class Redditor(MessageableMixin, RedditorListingMixin, FullnameMixin, RedditBase
         self.__dict__.update(other.__dict__)
         self._fetched = True
 
-    def _fetch_data(self):
-        name, fields, params = self._fetch_info()
-        path = API_PATH[name].format(**fields)
-        return self._reddit.request(method="GET", params=params, path=path)
-
     def _fetch_info(self):
         if hasattr(self, "_fullname"):
             self.name = self._fetch_username(self._fullname)