+++ praw/models/reddit/subreddit.py
@@ -830,11 +830,6 @@ class Subreddit(MessageableMixin, SubredditListingMixin, FullnameMixin, RedditBa
         self.__dict__.update(other.__dict__)
         self._fetched = True
 
-    def _fetch_data(self) -> dict:
-        name, fields, params = self._fetch_info()
-        path = API_PATH[name].format(**fields)
-        return self._reddit.request(method="GET", params=params, path=path)
-
     def _fetch_info(self):
         return "subreddit_about", {"subreddit": self}, None