+++ CHANGES.rst
@@ -3,8 +3,8 @@ Change Log
 
 PRAW follows `semantic versioning <http://semver.org/>`_.
 
-Unreleased
-----------
+7.7.1 (2023/07/11)
+------------------
 
 **Fixed**